Consolidation: test_isone.py
